firstRepeat doctest example
Add an example/doctest for firstRepeat which demonstrates
the expected behavior.

I thought this would be useful because I was initially
confused on the implementation. For a bit I thought
`firstRepeat` was supposed to find the first element
which has a repeat element somewhere in the list.
The example given demonstrates the important
distinction that we aren't looking for the first element
that has another of itself in the list, but that we're
looking for the first repeat element.